+++ src/app/trace.rs
@@ -1,13 +1,10 @@
 use async_trait::async_trait;
-use chrono::{DateTime, NaiveDateTime, Utc};
 use log::{debug, error, info, warn};
-use rand::distributions::{Distribution, Uniform};
 use std::convert::TryFrom;
-use std::time::{SystemTime, UNIX_EPOCH};
 use tokio::sync::mpsc::UnboundedSender;
 use tokio::time::{self, Duration, Instant};
 
-use crate::error::{ProtocolError, TaskError};
+use crate::error::{ProtocolError, TaskError, TraceError};
 use crate::protocol::Message;
 use crate::task::{Request, Task, TaskEvent, TaskOptions, TaskStatus};
 
@@ -50,14 +47,14 @@ impl<T> TracerTrait for Tracer<T>
 where
     T: Task,
 {
-    async fn trace(&mut self) -> Result<(), TaskError> {
+    async fn trace(&mut self) -> Result<(), TraceError> {
         if self.is_expired() {
             warn!(
                 "Task {}[{}] expired, discarding",
                 self.task.name(),
                 &self.task.request().id,
             );
-            return Err(TaskError::ExpirationError);
+            return Err(TraceError::ExpirationError);
         }
 
         self.event_tx
@@ -112,19 +109,20 @@ where
                             reason
                         );
                     }
-                    TaskError::TimeoutError => {
-                        error!(
-                            "Task {}[{}] timed out",
-                            self.task.name(),
-                            &self.task.request().id,
-                        );
-                    }
-                    _ => {
+                    TaskError::UnexpectedError(ref reason) => {
                         error!(
                             "Task {}[{}] failed with unexpected error: {}",
                             self.task.name(),
                             &self.task.request().id,
-                            e
+                            reason
+                        );
+                    }
+                    TaskError::TimeoutError => {
+                        error!(
+                            "Task {}[{}] timed out after {}s",
+                            self.task.name(),
+                            &self.task.request().id,
+                            duration.as_secs_f32(),
                         );
                     }
                 };
@@ -146,7 +144,7 @@ where
                             self.task.name(),
                             &self.task.request().id,
                         );
-                        return Err(e);
+                        return Err(TraceError::TaskError(e));
                     }
                     info!(
                         "Task {}[{}] retrying ({} / {})",
@@ -164,7 +162,7 @@ where
                     );
                 }
 
-                Err(TaskError::Retry)
+                Err(TraceError::Retry(self.task.retry_eta()))
             }
         }
     }
@@ -172,32 +170,6 @@ where
     async fn wait(&self) {
         if let Some(countdown) = self.task.request().countdown() {
             time::delay_for(countdown).await;
-        }
-    }
-
-    fn retry_eta(&self) -> Option<DateTime<Utc>> {
-        let retries = self.task.request().retries;
-        let delay_secs = std::cmp::min(
-            2u32.checked_pow(retries)
-                .unwrap_or_else(|| self.task.max_retry_delay().unwrap_or(3600)),
-            self.task.max_retry_delay().unwrap_or(3600),
-        );
-        let delay_secs = std::cmp::max(delay_secs, self.task.min_retry_delay().unwrap_or(0));
-        let between = Uniform::from(0..1000);
-        let mut rng = rand::thread_rng();
-        let delay_millis = between.sample(&mut rng);
-        match SystemTime::now().duration_since(UNIX_EPOCH) {
-            Ok(now) => {
-                let now_secs = now.as_secs() as u32;
-                let now_millis = now.subsec_millis();
-                let eta_secs = now_secs + delay_secs;
-                let eta_millis = now_millis + delay_millis;
-                Some(DateTime::<Utc>::from_utc(
-                    NaiveDateTime::from_timestamp(eta_secs as i64, eta_millis * 1000),
-                    Utc,
-                ))
-            }
-            Err(_) => None,
         }
     }
 
@@ -218,13 +190,10 @@ where
 pub(super) trait TracerTrait: Send + Sync {
     /// Wraps the execution of a task, catching and logging errors and then running
     /// the appropriate post-execution functions.
-    async fn trace(&mut self) -> Result<(), TaskError>;
+    async fn trace(&mut self) -> Result<(), TraceError>;
 
     /// Wait until the task is due.
     async fn wait(&self);
-
-    /// Get the ETA for a retry with exponential backoff.
-    fn retry_eta(&self) -> Option<DateTime<Utc>>;
 
     fn is_delayed(&self) -> bool;